requirements.txt
fastapi
uvicorn
line-bot-sdk
openai
pandas
yfinance
python-dateutil
openpyxl

